+++ src/Oro/Bundle/ImportExportBundle/Controller/ImportExportController.php
@@ -4,6 +4,8 @@ namespace Oro\Bundle\ImportExportBundle\Controller;
 
 use Oro\Bundle\ImportExportBundle\Async\ImportExportResultSummarizer;
 use Oro\Bundle\ImportExportBundle\Async\Topics;
+use Oro\Bundle\ImportExportBundle\Entity\ImportExportResult;
+use Oro\Bundle\ImportExportBundle\Exception\ImportExportExpiredException;
 use Oro\Bundle\ImportExportBundle\Exception\InvalidArgumentException;
 use Oro\Bundle\ImportExportBundle\File\FileManager;
 use Oro\Bundle\ImportExportBundle\Form\Model\ExportData;
@@ -13,12 +15,13 @@ use Oro\Bundle\ImportExportBundle\Form\Type\ExportType;
 use Oro\Bundle\ImportExportBundle\Form\Type\ImportType;
 use Oro\Bundle\ImportExportBundle\Handler\CsvFileHandler;
 use Oro\Bundle\ImportExportBundle\Handler\ExportHandler;
-use Oro\Bundle\ImportExportBundle\Handler\HttpImportHandler;
+use Oro\Bundle\ImportExportBundle\Handler\ImportHandler;
 use Oro\Bundle\ImportExportBundle\Job\JobExecutor;
 use Oro\Bundle\ImportExportBundle\Processor\ProcessorRegistry;
 use Oro\Bundle\MessageQueueBundle\Entity\Job;
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
 use Oro\Component\MessageQueue\Client\MessageProducerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
@@ -32,8 +35,11 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInt
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
+ * Controller for import/export actions
  * @SuppressWarnings(PHPMD.TooManyPublicMethods)
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
+ *
+ * Responsible for the import and export
  */
 class ImportExportController extends Controller
 {
@@ -235,7 +241,7 @@ class ImportExportController extends Controller
         $originFileName = $request->get('originFileName', null);
 
         $this->getMessageProducer()->send(
-            Topics::PRE_HTTP_IMPORT,
+            Topics::PRE_IMPORT,
             [
                 'fileName' => $fileName,
                 'process' => ProcessorRegistry::TYPE_IMPORT_VALIDATION,
@@ -269,7 +275,7 @@ class ImportExportController extends Controller
         $originFileName = $request->get('originFileName', null);
 
         $this->getMessageProducer()->send(
-            Topics::PRE_HTTP_IMPORT,
+            Topics::PRE_IMPORT,
             [
                 'fileName' => $fileName,
                 'process' => ProcessorRegistry::TYPE_IMPORT,
@@ -406,70 +412,56 @@ class ImportExportController extends Controller
             $this->getOptionsFromRequest($request)
         );
 
-        return $this->redirectToRoute('oro_importexport_export_download', ['fileName' => $result['file']]);
+        return $this->getExportHandler()->handleDownloadExportResult($result['file']);
     }
 
     /**
-     * @Route("/export/download/{fileName}", name="oro_importexport_export_download")
+     * @Route("/export/download/{jobId}", name="oro_importexport_export_download", requirements={"jobId"="\d+"})
+     * @ParamConverter("result", options={"mapping": {"jobId": "jobId"}})
      *
-     * @param string $fileName
+     * @param ImportExportResult $result
      *
      * @return Response
      */
-    public function downloadExportResultAction($fileName)
+    public function downloadExportResultAction(ImportExportResult $result)
     {
-        if (!$this->isGranted('oro_importexport_import')
-            && !$this->isGranted('oro_importexport_export')
-        ) {
+        if (!$this->isGranted('VIEW', $result)) {
             throw new AccessDeniedException('Insufficient permission');
         }
 
-        return $this->getExportHandler()->handleDownloadExportResult($fileName);
+        if ($result->isExpired()) {
+            throw new ImportExportExpiredException();
+        }
+
+        return $this->getExportHandler()->handleDownloadExportResult($result->getFilename());
     }
 
     /**
-     * @Route("/import_export/error/{jobCode}.log", name="oro_importexport_error_log")
+     * @Route(
+     *     "/import_export/job-error-log/{jobId}.log",
+     *     name="oro_importexport_job_error_log",
+     *     requirements={"jobId"="\d+"}
+     * )
+     * @ParamConverter("result", options={"mapping": {"jobId": "jobId"}})
      *
-     * @param string $jobCode
+     * @param ImportExportResult $result
      * @return Response
-     * @throws AccessDeniedException
      */
-    public function errorLogAction($jobCode)
+    public function importExportJobErrorLogAction(ImportExportResult $result)
     {
-        if (!$this->isGranted('oro_importexport_import')
-            && !$this->isGranted('oro_importexport_export')
-        ) {
+        if (!$this->isGranted('VIEW', $result)) {
             throw new AccessDeniedException('Insufficient permission');
         }
 
-        $jobExecutor = $this->getJobExecutor();
-        $errors      = array_merge(
-            $jobExecutor->getJobFailureExceptions($jobCode),
-            $jobExecutor->getJobErrors($jobCode)
-        );
-        $content     = implode("\r\n", $errors);
-
-        return new Response($content, 200, ['Content-Type' => 'text/x-log']);
-    }
-
-    /**
-     * @Route("/import_export/job-error-log/{jobId}.log", name="oro_importexport_job_error_log")
-     *
-     * @param $jobId
-     * @return Response
-     */
-    public function importExportJobErrorLogAction($jobId)
-    {
-        if (!$this->isGranted('oro_importexport_import')
-            && !$this->isGranted('oro_importexport_export')
-        ) {
-            throw new AccessDeniedException('Insufficient permission');
+        if ($result->isExpired()) {
+            throw new ImportExportExpiredException();
         }
 
-        $job = $this->getDoctrine()->getManagerForClass(Job::class)->getRepository(Job::class)->find($jobId);
+        $jobRepository = $this->getDoctrine()->getManagerForClass(Job::class)->getRepository(Job::class);
+        $job = $jobRepository->find($result->getJobId());
 
         if (!$job) {
-            throw new NotFoundHttpException(sprintf('Job %s not found', $jobId));
+            throw new NotFoundHttpException(sprintf('Job %s not found', $result->getJobId()));
         }
 
         $content = $this->getImportExportResultSummarizer()->getErrorLog($job);
@@ -478,11 +470,11 @@ class ImportExportController extends Controller
     }
 
     /**
-     * @return HttpImportHandler
+     * @return ImportHandler
      */
     protected function getImportHandler()
     {
-        return $this->get('oro_importexport.handler.import.http');
+        return $this->get('oro_importexport.handler.import');
     }
 
     /**